Renamed value() methods that return strings to text(), should be less confusing and will allow combo-box widgets that return both numbers and text. Put in back-compatability value() method on Input.
Removed value() from the base class. Cast Widget to Button to get it's value.
Added "make_exec_return(bool)" method to replace other use of value.

Removed clamping from Valuator. Please put this in the callback or subclass.

Fixed the TextEditor going into an infinite loop on bad utf-8. Still
needs a lot of work!

Changes to the TextEditor so it is a drop-in substitute for Input widget,
by making it create the TextBuffer itself and adding text() and some other
methods.

Made prototype for filename_list sorting functions const-correct.

Merged filename_normalize into filename_absolute by adding extra optional
argument for the current directory. Made it return the length of resulting
string.

Added the dx() etc methods to base Symbol class, made them call inset().
This will allow images to be used for boxes.

Made Preferences functions return bool. Moved internal structurs to src file.


git-svn-id: file:///fltk/svn/trunk@4843 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/win32/run.cxx
@@ -1207,7 +1207,7 @@ static unsigned long shiftflags(bool ignorealt=false) {
 
 static unsigned recent_keysym = 0;
 
-static bool mouse_event(Window *window, int what, int button,
+static bool mouse_event(Window *window, int what, unsigned button,
 			WPARAM wParam, LPARAM lParam)
 {
   xmousewin = window;
@@ -1253,7 +1253,7 @@ static bool mouse_event(Window *window, int what, int button,
   switch (what) {
   case 1: // double-click
     // This is not detecting triple-clicks, does anybody know how to fix?
-    if (e_is_click==(unsigned)button) {e_clicks++; goto J1;}
+    if (e_is_click == button) {e_clicks++; goto J1;}
   case 0: // single-click
     e_clicks = 0;
   J1:
